+++ services/question-service/app/routers/__init__.py
@@ -1,0 +1,5 @@
+# Router imports for the question service
+from .questions import router as questions_router
+from .metadata import router as metadata_router
+
+__all__ = ["questions_router", "metadata_router"]
